Initial draft of new NCAE collection interface
This commit introduces an initial draft of the new NCAE interface, which offers
custom modules, plugins and roles to greatly simplify the development of modules
for NCAE. It is supposed to hide technical implementation details from the user
and allows an easier and more standardized approach for module development.

At the state of being, this does not include the full API of the NCAE core yet.
Instead it focuses on the most important endpoints that are required when
setting up AWX-based services in a module. Further functionality will be added
in upcoming merge requests to gradually implement a new API.

This interface is NOT stable at the moment and should not be used.